Rename service watcher abstract class to services package
