+++ app/src/main/java/com/amaze/filemanager/asynchronous/services/ExtractService.java
@@ -38,7 +38,6 @@ import com.amaze.filemanager.filesystem.FileUtil;
 import com.amaze.filemanager.ui.notifications.NotificationConstants;
 import com.amaze.filemanager.utils.CopyDataParcelable;
 import com.amaze.filemanager.utils.ProgressHandler;
-import com.amaze.filemanager.utils.ServiceWatcherProgressAbstract;
 import com.amaze.filemanager.utils.ServiceWatcherUtil;
 import com.amaze.filemanager.utils.application.AppConfig;
 import com.amaze.filemanager.utils.files.GenericCopyUtil;
